V0.5.1 — NPM module update
Updated file structure for publishing to npm
