+++ .gitignore
@@ -9,7 +9,8 @@
 /coverage
 
 # production
-/build
+/examples/dist
+/dist
 
 # misc
 .DS_Store
