+++ examples/src/index.html
@@ -3,11 +3,9 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
+  <title>Shooting Stars Background</title>
 </head>
-<body style="margin: 0;">
-  <div id="root" style="width: 100vw; height: 100vh; margin: 0;">
-
-  </div>
+<body style="margin: 0">
+  <div style="width: 100vw; height: 100vh;" id="root"></div>
 </body>
 </html>
